+++ .gitignore
@@ -1,5 +1,8 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
+# vs code
+.vscode/
+
 # dependencies
 /node_modules
 /.pnp
@@ -35,3 +38,4 @@ yarn-error.log*
 
 # vercel
 .vercel
+
