+++ components/klaudsolcms/dropdown/SidebarFooterIcon.js
@@ -14,7 +14,9 @@ const SidebarFooterIcon = ({title}) => {
     const onLogout = (evt) => {
         evt.preventDefault();   
         const callback = async () => {
-          await slsFetch('/api/auth/logout');
+          await slsFetch('/api/session',{
+            method: "DELETE"
+          });
           dispatch({type: RESET_CLIENT_SESSION});
           router.push('/');
         };
